Cleanup: move Job render call to Job class
